backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
telethon>=1.30.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
watchfiles>=0.19.0